docs: seperate UserError file
