+++ FOFMAP/Domain/UseCase/UserMatchRecordUseCase.swift
@@ -7,20 +7,6 @@
 
 import Foundation
 
-enum UserError: LocalizedError {
-    case noExistUser
-    case noExistMatchRecord
-    
-    var errorDescription: String? {
-        switch self {
-        case .noExistUser:
-            return "존재하지 않는 유저입니다."
-        case .noExistMatchRecord:
-            return "매치기록이 존재하지 않습니다."
-        }
-    }
-}
-
 struct UserMatchRecordUseCase: DetailFetchable {
     typealias T = [MatchDesc]
